backend/requirements.txt
asgiref==3.5.2
bleach==5.0.1
boto3==1.24.87
botocore==1.27.87
certifi==2022.9.24
cffi==1.15.1
chardet==5.0.0
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.2
django-colorfield==0.7.2
django-cors-headers==3.13.0
django-ebhealthcheck==2.0.2
django-environ==0.9.0
django-stdimage==6.0.1
django-extensions==3.2.1
django-filter==22.1
django-polymorphic==3.1.0
django-rest-polymorphic==0.1.10
django-ses==3.2.0
django-storages==1.13.1
djangorestframework==3.14.0
djangorestframework-camel-case==1.3.0
docutils==0.19
drf-api-logger==1.1.11
drf-nested-routers==0.93.4
et-xmlfile==1.1.0
idna==3.4
jmespath==1.0.1
josepy==1.13.0
mozilla-django-oidc==2.0.0
openpyxl==3.0.10
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==22.0.0
python-dateutil==2.8.2
pytz==2022.1
requests==2.28.1
s3transfer==0.6.0
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
webencodings==0.5.1
